+++ test/list/read-list-info.js
@@ -153,7 +153,7 @@ describe('read M1F1-int12WE-AFsp-NEW-TAGS.wav and write to new file', function()
         assert.equal(wav.LIST[0]["subChunks"][5]["value"], "CopyAudio");
     });
     it("wav.LIST[0]['chunkSize'] == wav.getLISTSize_()", function() {
-        assert.equal(wav.LIST[0]["chunkSize"], wav.getLISTSize_() - 8);
+        assert.equal(wav.LIST[0]["chunkSize"], wav.getLISTBytes_().length - 8);
     });
 
     // Reading tags from the file written with WaveFile
@@ -218,7 +218,7 @@ describe('read M1F1-int12WE-AFsp-NEW-TAGS.wav and write to new file', function()
     it("wav.LIST[0]['chunkSize'] == wav.getLISTSize_()",
         function() {
         assert.equal(
-            wav2.LIST[0]["chunkSize"], wav.getLISTSize_() - 8);
+            wav2.LIST[0]["chunkSize"], wav.getLISTBytes_().length - 8);
     });
 
     // Other tests
@@ -452,7 +452,7 @@ describe('read Audacity-16bit.wav and write to new file', function() {
         assert.equal(wav2.chunkSize + 8, fileSizeInBytes2);
     });
     it("wav.LIST[0]['chunkSize'] == wav.getLISTSize_()", function() {
-        assert.equal(wav2.LIST[0]["chunkSize"], wav2.getLISTSize_() - 8);
+        assert.equal(wav2.LIST[0]["chunkSize"], wav2.getLISTBytes_().length - 8);
     });
     it("wav2.cue should be == wav.cue", function() {
         assert.deepEqual(wav2.cue, wav.cue);